+++ src/provider/index.ts
@@ -1,4 +1,7 @@
-import { IntegrationInstance } from '@jupiterone/integration-sdk-core';
+import {
+  IntegrationInstance,
+  IntegrationLogger,
+} from '@jupiterone/integration-sdk-core';
 
 import { GitlabIntegrationConfig } from '../types';
 import { GitlabClient } from './GitlabClient';
@@ -8,10 +11,18 @@ type GitlabClientConfig = {
   personalToken: string;
 };
 
-export function createGitlabClient(config: GitlabClientConfig): GitlabClient {
-  return new GitlabClient(config.baseUrl, config.personalToken);
+let client: GitlabClient;
+export function createGitlabClient(
+  config: GitlabClientConfig,
+  logger: IntegrationLogger,
+): GitlabClient {
+  if (!client) {
+    client = new GitlabClient(config.baseUrl, config.personalToken, logger);
+  }
+  return client;
 }
 
 export type ClientCreator = (
   client: IntegrationInstance<GitlabIntegrationConfig>,
+  logger: IntegrationLogger,
 ) => GitlabClient;