+++ src/steps/build-user-approved-merge-request-relationships/index.ts
@@ -25,8 +25,9 @@ export function createStep(
     async executionHandler({
       jobState,
       instance,
+      logger,
     }: IntegrationStepExecutionContext<GitlabIntegrationConfig>): Promise<void> {
-      const client = clientCreator(instance);
+      const client = clientCreator(instance, logger);
       const userIdMap = await createUserIdMap(jobState);
 
       await jobState.iterateEntities(
@@ -73,7 +74,9 @@ async function createUserIdMap(
   return userIdMap;
 }
 
-export default createStep((instance) => createGitlabClient(instance.config));
+export default createStep((instance, logger) =>
+  createGitlabClient(instance.config, logger),
+);
 
 export function createUserApprovedPrRelationship(
   user: Entity,